+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 ## 0.4.0.0
 
+- Remvoe the `TransmissionTree` and `SampleTree` data types because there is
+  already the `EpidemicTree` and `ReconstructedTree` in the
+  `Epidemic.Types.Events` which should be used preferentially.
 - Remove dependency on `trifecta` since this functionality is not necessary.
 
 ## 0.3.0.0